+++ parity-crypto/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog].
 [Keep a Changelog]: http://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
+
+## [0.7.0] - 2021-01-05
 ### Breaking
 - Bump `rust-secp256k1` to v0.19, always allow zero signatures. [#438](https://github.com/paritytech/parity-common/pull/438)
 - Updated `rlp` to 0.5. [#463](https://github.com/paritytech/parity-common/pull/463)